bert-mlm port to keras-core with tf backend (#843)
* bert-mlm port to keras-core with tf backend

* added MaskedTextGenerator callback, fixed imports

* lint fixes

---------

+++ keras_core/initializers/constant_initializers.py
@@ -29,11 +29,13 @@ class Constant(Initializer):
     """
 
     def __init__(self, value=0.0):
-        self.value = float(value)
+        self.value = value
 
     def __call__(self, shape, dtype=None):
         dtype = standardize_dtype(dtype)
-        return self.value * ops.ones(shape=shape, dtype=dtype)
+        return ops.cast(self.value, dtype=dtype) * ops.ones(
+            shape=shape, dtype=dtype
+        )
 
     def get_config(self):
         return {"value": self.value}
